Added Flat Integrator

+++ include/ray.h
@@ -6,12 +6,12 @@ class Ray {
  private:
   point3 origin;
   vec3 direction;
-  float i;
-  float j;
+  double i;
+  double j;
 
  public:
-  Ray(const point3 &_origin, const vec3 &_direction, const int &_i,
-      const int &_j)
+  Ray(const point3 &_origin, const vec3 &_direction, const double &_i,
+      const double &_j)
       : origin(_origin), direction(_direction), i(_i), j(_j) {
     /* Empty */
   }
@@ -20,8 +20,8 @@ class Ray {
 
   point3 getOrigin() const { return origin; }
   vec3 getDirection() const { return direction; }
-  int getI() const { return i; }
-  int getJ() const { return j; }
+  double getI() const { return i; }
+  double getJ() const { return j; }
 
   friend std::ostream &operator<<(std::ostream &os, const Ray &t) {
     os << "[o=" << t.origin << ", d=" << t.direction << "]";
